requirements.txt
# LLM & RAG
groq>=0.4.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
regex>=2023.0.0

# API & UI
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
